backend/requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.23
requests==2.26.0
Werkzeug==2.2.2
